Feat initial commit for user-tenant-service database migration using Flyway (#57)
User-tenant-service database migration using Flyway (Versioning +
Database migration)

## Description

Flyway is a database migration tool that helps you manage and apply
database changes in a consistent and reliable way. It is easy to use and
integrate with your existing tools and frameworks.

Flyway migrate the database as well as keep track of versioning using a
flyway_schema_history table ,
it detect the version and changes using flyway_schema_history table and
apply the migration if required or else if there is no new changes or
update is required it simply skip the migration with message " database
up to date hence no migration is required."

Versioning also help us to keep track of changes in database
periodically based on buiseneness requirement.

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Intermediate change (work in progress)

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

- [x] Test A
- [x] Test B

## Checklist:

- [ ] Performed a self-review of my own code
- [ ] npm test passes on your machine
- [ ] New tests added or existing tests modified to cover all changes
- [ ] Code conforms with the style guide
- [ ] API Documentation in code was updated
- [ ] Any dependent changes have been merged and published in downstream
modules